refactor: move hasRelativeItems fn
